Getting multiple sensor data form Postgres into python
Two dictionaries (measurements/locations) hold the measurement and location data respectively.
To access them first requires the sensor_id, which is used as a key in a lookup dictionary to get the local key.
Use the local key to search for what sensor you want to look at in either the measurement or location dictionary.